Add role on all method

+++ src/Http/Controllers/PendaftaranController.php
@@ -10,10 +10,12 @@ use Carbon\Carbon;
 use Bantenprov\Pendaftaran\Models\Bantenprov\Pendaftaran\Pendaftaran;
 use Bantenprov\Kegiatan\Models\Bantenprov\Kegiatan\Kegiatan;
 use Bantenprov\Sekolah\Models\Bantenprov\Sekolah\Sekolah;
+use Bantenprov\Sekolah\Models\Bantenprov\Sekolah\AdminSekolah;
 use App\User;
 
 /* Etc */
 use Validator;
+use Auth;
 
 /**
  * The PendaftaranController class.
@@ -32,13 +34,15 @@ class PendaftaranController extends Controller
     protected $pendaftaran;
     protected $user;
     protected $sekolah;
+    protected $admin_sekolah;
 
-    public function __construct(Pendaftaran $pendaftaran, Kegiatan $kegiatan, User $user, Sekolah $sekolah)
+    public function __construct(Pendaftaran $pendaftaran, Kegiatan $kegiatan, User $user, Sekolah $sekolah, AdminSekolah $admin_sekolah)
     {
         $this->pendaftaran      = $pendaftaran;
         $this->kegiatanModel    = $kegiatan;
         $this->user             = $user;
         $this->sekolah          = $sekolah;
+        $this->admin_sekolah    = $admin_sekolah;
     }
 
     /**
@@ -48,19 +52,38 @@ class PendaftaranController extends Controller
      */
     public function index(Request $request)
     {
+        $admin_sekolah = $this->admin_sekolah->where('admin_sekolah_id', Auth::user()->id)->first();
+
+        if(is_null($admin_sekolah) && $this->checkRole(['superadministrator']) === false){
+            $response = [];
+            return response()->json($response)
+            ->header('Access-Control-Allow-Origin', '*')
+            ->header('Access-Control-Allow-Methods', 'GET');
+        }
+
+
         if (request()->has('sort')) {
             list($sortCol, $sortDir) = explode('|', request()->sort);
 
-            $query = $this->pendaftaran->with('kegiatan')->with('user')->orderBy($sortCol, $sortDir);
+            if($this->checkRole(['superadministrator'])){
+                $query = $this->pendaftaran->with('kegiatan')->with('user')->orderBy($sortCol, $sortDir);
+            }else{
+                $query = $this->pendaftaran->where('sekolah_id', $admin_sekolah->sekolah_id)->with('kegiatan')->with('user')->orderBy($sortCol, $sortDir);
+            }
+
         } else {
-            $query = $this->pendaftaran->with('kegiatan')->with('user')->orderBy('id', 'asc');
+            if($this->checkRole(['superadministrator'])){
+                $query = $this->pendaftaran->with('kegiatan')->with('user')->orderBy($sortCol, $sortDir);
+            }else{
+                $query = $this->pendaftaran->where('sekolah_id', $admin_sekolah->sekolah_id)->with('kegiatan')->with('user')->orderBy('id', 'asc');
+            }
+
         }
 
         if ($request->exists('filter')) {
             $query->where(function($q) use($request) {
                 $value = "%{$request->filter}%";
-                $q->where('tanggal_pendaftaran', 'like', $value)
-                    ->orWhere('tanggal_pendaftaran', 'like', $value);
+                $q->where('tanggal_pendaftaran', 'like', $value);
             });
         }
 
@@ -79,10 +102,17 @@ class PendaftaranController extends Controller
      */
     public function create()
     {
+        $admin_sekolah = $this->admin_sekolah->where('admin_sekolah_id', Auth::user()->id)->first();
+
         $response = [];
 
         $kegiatan       = $this->kegiatanModel->all();
-        $sekolahs       = $this->sekolah->all();
+        if($this->checkRole(['superadministrator'])){
+            $sekolahs       = $this->sekolah->all();
+        }else{
+            $sekolahs       = $this->sekolah->where('id',$admin_sekolah->sekolah_id)->get();
+        }
+
         $users_special  = $this->user->all();
         $users_standar  = $this->user->find(\Auth::User()->id);
         $current_user   = \Auth::User();
@@ -126,6 +156,8 @@ class PendaftaranController extends Controller
      */
     public function store(Request $request)
     {
+        $admin_sekolah = $this->admin_sekolah->where('admin_sekolah_id', Auth::user()->id)->first();
+
         $pendaftaran        = $this->pendaftaran;
         $current_user_id    = $request->user_id;
 
@@ -135,6 +167,14 @@ class PendaftaranController extends Controller
             'tanggal_pendaftaran'   => 'required',
             'sekolah_id'            => 'required',
         ]);
+
+        if($admin_sekolah->sekolah_id != $request->sekolah_id && $this->checkRole(['superadministrator']) === false){
+            $response['error']      = true;
+            $response['message']    = 'Terjadi kesalahan, mohon ulangi lagi pengisian data yang benar.';
+            $response['status']     = true;
+
+            return response()->json($response);
+        }
 
         if ($validator->fails()) {
             $error      = true;
@@ -166,7 +206,15 @@ class PendaftaranController extends Controller
      */
     public function show($id)
     {
-        $pendaftaran = $this->pendaftaran->findOrFail($id);
+        $admin_sekolah = $this->admin_sekolah->where('admin_sekolah_id', Auth::user()->id)->first();
+
+        if($this->checkRole(['superadministrator'])){
+            $pendaftaran = $this->pendaftaran->findOrFail($id);
+        }else{
+            $pendaftaran = $this->pendaftaran->where('sekolah_id', $admin_sekolah->sekolah_id)->findOrFail($id);
+        }
+
+
 
         $response['pendaftaran']    = $pendaftaran;
         $response['kegiatan']       = $pendaftaran->kegiatan;
@@ -185,8 +233,14 @@ class PendaftaranController extends Controller
      */
     public function edit($id)
     {
+        $admin_sekolah = $this->admin_sekolah->where('admin_sekolah_id', Auth::user()->id)->first();
 
-        $pendaftaran = $this->pendaftaran->with(['kegiatan', 'sekolah', 'user' ])->findOrFail($id);
+        if($this->checkRole(['superadministrator'])){
+            $pendaftaran = $this->pendaftaran->with(['kegiatan', 'sekolah', 'user' ])->findOrFail($id);
+        }else{
+            $pendaftaran = $this->pendaftaran->where('sekolah_id', $admin_sekolah->sekolah_id)->with(['kegiatan', 'sekolah', 'user' ])->findOrFail($id);
+        }
+
 
         $response['pendaftaran']['user']     = array_add($pendaftaran->user, 'label', $pendaftaran->user->name);
 
@@ -212,21 +266,34 @@ class PendaftaranController extends Controller
      */
     public function update(Request $request, $id)
     {
-         $pendaftaran = $this->pendaftaran->with(['kegiatan','sekolah','user'])->findOrFail($id);
+        $admin_sekolah = $this->admin_sekolah->where('admin_sekolah_id', Auth::user()->id)->first();
 
-            $validator = Validator::make($request->all(), [
-                'user_id'               => "required|exists:{$this->user->getTable()},id",
-                'kegiatan_id'           => 'required',
-                'tanggal_pendaftaran'   => 'required',
-                'sekolah_id'            => 'required',
-            ]);
+        if($this->checkRole(['superadministrator'])){
+            $pendaftaran = $this->pendaftaran->with(['kegiatan', 'sekolah', 'user' ])->findOrFail($id);
+        }else{
+            $pendaftaran = $this->pendaftaran->where('sekolah_id', $admin_sekolah->sekolah_id)->with(['kegiatan', 'sekolah', 'user' ])->findOrFail($id);
+
+            if(is_null($pendaftaran)){
+                $response['error']      = true;
+                $response['message']    = 'Terjadi kesalahan saat update data.';
+                $response['status']     = true;
+
+                return response()->json($response);
+            }
+        }
+
+
+        $validator = Validator::make($request->all(), [
+            'kegiatan_id'           => 'required',
+            'tanggal_pendaftaran'   => 'required',
+            'sekolah_id'            => 'required',
+        ]);
 
         if ($validator->fails()) {
             $error      = true;
             $message    = $validator->errors()->first();
 
         } else {
-            $pendaftaran->user_id               = $request->input('user_id');
             $pendaftaran->kegiatan_id           = $request->input('kegiatan_id');
             $pendaftaran->tanggal_pendaftaran   = $request->input('tanggal_pendaftaran');
             $pendaftaran->sekolah_id            = $request->input('sekolah_id');
@@ -251,6 +318,11 @@ class PendaftaranController extends Controller
      */
     public function destroy($id)
     {
+        if(!$this->checkRole(['superadministrator'])){
+            $response['status'] = false;
+            return json_encode($response);
+        }
+
         $pendaftaran = $this->pendaftaran->findOrFail($id);
 
         if ($pendaftaran->delete()) {
@@ -260,5 +332,10 @@ class PendaftaranController extends Controller
         }
 
         return json_encode($response);
+    }
+
+    protected function checkRole($role = array())
+    {
+        return Auth::user()->hasRole($role);
     }
 }
